+++ src/components/calcite-action/calcite-action.e2e.ts
@@ -6,36 +6,29 @@ describe("calcite-action", () => {
 
   it("honors hidden attribute", async () => hidden("calcite-action"));
 
-  it("should not have text container", async () => {
+  it("should have visible text when text is enabled", async () => {
     const page = await newE2EPage();
+    await page.setContent(`<calcite-action text="hello world" text-enabled></calcite-action>`);
 
-    await page.setContent(`<calcite-action text="hello world"></calcite-action>`).catch((error) => {
-      console.error(error);
-    });
+    const textContainer = await page.find("calcite-action >>> .text-container");
+    const isVisible = await textContainer.isVisible();
 
-    const textcontainer = await page.find("calcite-action >>> .text-container");
-    expect(textcontainer).toBeNull();
+    expect(isVisible).toBe(true);
   });
 
-  it("should have text container", async () => {
+  it("should not have visible text when text is not enabled", async () => {
     const page = await newE2EPage();
+    await page.setContent(`<calcite-action text="hello world"></calcite-action>`);
 
-    await page
-      .setContent(`<calcite-action text="hello world" text-display="visible"></calcite-action>`)
-      .catch((error) => {
-        console.error(error);
-      });
+    const textContainer = await page.find("calcite-action >>> .text-container");
+    const isVisible = await textContainer.isVisible();
 
-    const textcontainer = await page.find("calcite-action >>> .text-container");
-    expect(textcontainer).not.toBeNull();
+    expect(isVisible).toBe(false);
   });
 
   it("should have icon container", async () => {
     const page = await newE2EPage();
-
-    await page.setContent(`<calcite-action><svg><path></svg></calcite-action>`).catch((error) => {
-      console.error(error);
-    });
+    await page.setContent(`<calcite-action><svg><path></svg></calcite-action>`);
 
     const iconContainer = await page.find("calcite-action >>> .icon-container");
     expect(iconContainer).not.toBeNull();
@@ -43,21 +36,15 @@ describe("calcite-action", () => {
 
   it("should always have icon container", async () => {
     const page = await newE2EPage();
-
-    await page.setContent(`<calcite-action></calcite-action>`).catch((error) => {
-      console.error(error);
-    });
+    await page.setContent(`<calcite-action></calcite-action>`);
 
     const iconContainer = await page.find("calcite-action >>> .icon-container");
     expect(iconContainer).not.toBeNull();
   });
 
-  it("should have fallback label", async () => {
+  it("should use text prop for a11y attributes when text is not enabled", async () => {
     const page = await newE2EPage();
-
-    await page.setContent(`<calcite-action text="hello world"></calcite-action>`).catch((error) => {
-      console.error(error);
-    });
+    await page.setContent(`<calcite-action text="hello world"></calcite-action>`);
 
     const button = await page.find("calcite-action >>> .button");
     expect(button.getAttribute("title")).toBe("hello world");
@@ -66,10 +53,7 @@ describe("calcite-action", () => {
 
   it("should have label", async () => {
     const page = await newE2EPage();
-
-    await page.setContent(`<calcite-action text="hello world" label="hi"></calcite-action>`).catch((error) => {
-      console.error(error);
-    });
+    await page.setContent(`<calcite-action text="hello world" label="hi"></calcite-action>`);
 
     const button = await page.find("calcite-action >>> .button");
     expect(button.getAttribute("title")).toBe("hi");
@@ -78,30 +62,22 @@ describe("calcite-action", () => {
 
   it("should be disabled", async () => {
     const page = await newE2EPage();
-
-    await page.setContent(`<calcite-action disabled></calcite-action>`).catch((error) => {
-      console.error(error);
-    });
+    await page.setContent(`<calcite-action disabled></calcite-action>`);
 
     const button = await page.find("calcite-action >>> .button");
-
     expect(button).toHaveAttribute("disabled");
   });
 
   it("should have appearance=solid", async () => {
     const page = await newE2EPage();
-
-    await page.setContent(`<calcite-action text="hello world"></calcite-action>`).catch((error) => {
-      console.error(error);
-    });
+    await page.setContent(`<calcite-action text="hello world"></calcite-action>`);
 
     const action = await page.find("calcite-action");
     expect(action.getAttribute("appearance")).toBe("solid");
   });
 
   it("should be accessible", async () => {
-    await accessible(`<calcite-action text="hello world" label="hi"></calcite-action>`);
-
-    await accessible(`<calcite-action text="hello world" label="hi" disabled text-display="visible"></calcite-action>`);
+    await accessible(`<calcite-action text="hello world"></calcite-action>`);
+    await accessible(`<calcite-action text="hello world" disabled text-enabled></calcite-action>`);
   });
 });