+++ src/components/calcite-block-section/calcite-block-section.tsx
@@ -1,9 +1,8 @@
 import { Component, Element, Event, EventEmitter, Prop, h } from "@stencil/core";
 
-import { caretDown16, caretLeft16, caretRight16 } from "@esri/calcite-ui-icons";
 import { getElementDir } from "../utils/dom";
-import { CSS, TEXT } from "./resources";
-import CalciteIcon from "../utils/CalciteIcon";
+import { CSS, ICONS, TEXT } from "./resources";
+import { guid } from "../utils/guid";
 import classnames from "classnames";
 import { CalciteBlockSectionToggleDisplay } from "../interfaces";
 
@@ -68,6 +67,8 @@ export class CalciteBlockSection {
   @Element()
   el: HTMLCalciteBlockSectionElement;
 
+  guid = `calcite-block-section-${guid()}`;
+
   // --------------------------------------------------------------------------
   //
   //  Events
@@ -97,16 +98,30 @@ export class CalciteBlockSection {
   // --------------------------------------------------------------------------
 
   render() {
-    const { el, open, text, textCollapse, textExpand, toggleDisplay } = this;
+    const { el, guid: id, open, text, textCollapse, textExpand, toggleDisplay } = this;
     const dir = getElementDir(el);
-    const arrowIcon = open ? caretDown16 : dir === "rtl" ? caretLeft16 : caretRight16;
+    const arrowIcon = open
+      ? ICONS.menuOpen
+      : dir === "rtl"
+      ? ICONS.menuClosedLeft
+      : ICONS.menuClosedRight;
     const toggleLabel = open ? textCollapse : textExpand;
+    const labelId = `${id}__label`;
 
     const headerNode =
       toggleDisplay === "switch" ? (
-        <label aria-label={toggleLabel} class={classnames(CSS.toggle, CSS.toggleSwitch)}>
+        <label
+          aria-label={toggleLabel}
+          class={classnames(CSS.toggle, CSS.toggleSwitch)}
+          id={labelId}
+        >
           {text}
-          <calcite-switch switched={open} onChange={this.toggleSection} scale="s" />
+          <calcite-switch
+            aria-labelledby={labelId}
+            switched={open}
+            onChange={this.toggleSection}
+            scale="s"
+          />
         </label>
       ) : (
         <calcite-action
@@ -114,10 +129,10 @@ export class CalciteBlockSection {
           class={CSS.toggle}
           onClick={this.toggleSection}
           text={text}
-          textDisplay="visible"
+          textEnabled={true}
           compact
         >
-          <CalciteIcon size="16" path={arrowIcon} />
+          <calcite-icon scale="s" icon={arrowIcon} />
         </calcite-action>
       );