+++ src/components/calcite-flow-item/calcite-flow-item.tsx
@@ -1,20 +1,17 @@
 import { Component, Element, Event, EventEmitter, Host, Prop, h } from "@stencil/core";
-import { VNode } from "@stencil/core/dist/declarations";
-
-import { chevronLeft16F, chevronRight16F, ellipsis16 } from "@esri/calcite-ui-icons";
-
-import { getElementDir } from "../utils/dom";
-
+import { VNode } from "@stencil/core/internal";
+import { focusElement, getElementDir } from "../utils/dom";
 import classnames from "classnames";
+import { BLACKLISTED_MENU_ACTIONS_COMPONENTS, CSS, ICONS, SLOTS, TEXT } from "./resources";
+import { getRoundRobinIndex } from "../utils/array";
+import { CalciteScale, CalciteTheme } from "../interfaces";
+import { CSS_UTILITY } from "../utils/resources";
 
-import { BLACKLISTED_MENU_ACTIONS_COMPONENTS, CSS, TEXT } from "./resources";
-import CalciteIcon from "../utils/CalciteIcon";
-
-import { CalciteTheme } from "../interfaces";
+const SUPPORTED_ARROW_KEYS = ["ArrowUp", "ArrowDown"];
 
 /**
- * @slot menu-actions - A slot for adding `calcite-actions` to a menu under the `...` in the header. These actions are displayed when the menu is open.
- * @slot footer-actions - A slot for adding `calcite-actions` to the footer.
+ * @slot menu-actions - A slot for adding `calcite-action`s to a menu under the `...` in the header. These actions are displayed when the menu is open.
+ * @slot footer-actions - A slot for adding `calcite-button`s to the footer.
  * @slot - A slot for adding content to the flow item.
  */
 @Component({
@@ -30,9 +27,19 @@ export class CalciteFlowItem {
   // --------------------------------------------------------------------------
 
   /**
+   * When provided, this method will be called before it is removed from the parent flow.
+   */
+  @Prop() beforeBack?: () => Promise<void>;
+
+  /**
    * When true, disabled prevents interaction. This state shows items with lower opacity/grayed.
    */
   @Prop({ reflect: true }) disabled = false;
+
+  /**
+   * Specifies the maxiumum height of the panel that this wraps.
+   */
+  @Prop({ reflect: true }) heightScale: CalciteScale;
 
   /**
    * Heading text.
@@ -53,6 +60,11 @@ export class CalciteFlowItem {
    * Shows a back button in the header.
    */
   @Prop() showBackButton = false;
+
+  /**
+   * Summary text. A description displayed underneath the heading.
+   */
+  @Prop() summary?: string;
 
   /**
    * 'Back' text string.
@@ -76,14 +88,6 @@ export class CalciteFlowItem {
 
   // --------------------------------------------------------------------------
   //
-  //  Private Properties
-  //
-  // --------------------------------------------------------------------------
-
-  @Element() el: HTMLCalciteFlowItemElement;
-
-  // --------------------------------------------------------------------------
-  //
   //  Events
   //
   // --------------------------------------------------------------------------
@@ -96,9 +100,25 @@ export class CalciteFlowItem {
 
   // --------------------------------------------------------------------------
   //
+  //  Private Properties
+  //
+  // --------------------------------------------------------------------------
+
+  @Element() el: HTMLCalciteFlowItemElement;
+
+  // --------------------------------------------------------------------------
+  //
   //  Private Methods
   //
   // --------------------------------------------------------------------------
+
+  queryActions(): HTMLCalciteActionElement[] {
+    return Array.from(this.el.querySelectorAll(`[slot=${SLOTS.menuActions}] calcite-action`));
+  }
+
+  isValidKey(key: string, supportedKeys: string[]): boolean {
+    return !!supportedKeys.find((k) => k === key);
+  }
 
   toggleMenuOpen = (): void => {
     this.menuOpen = !this.menuOpen;
@@ -106,6 +126,76 @@ export class CalciteFlowItem {
 
   backButtonClick = (): void => {
     this.calciteFlowItemBackClick.emit();
+  };
+
+  menuButtonKeyDown = (event: KeyboardEvent): void => {
+    const { key } = event;
+    const { menuOpen } = this;
+
+    if (!this.isValidKey(key, SUPPORTED_ARROW_KEYS)) {
+      return;
+    }
+
+    const actions = this.queryActions();
+    const { length } = actions;
+
+    if (!length) {
+      return;
+    }
+
+    event.preventDefault();
+
+    if (!menuOpen) {
+      this.menuOpen = true;
+    }
+
+    if (key === "ArrowUp") {
+      const lastAction = actions[length - 1];
+      focusElement(lastAction);
+    }
+
+    if (key === "ArrowDown") {
+      const firstAction = actions[0];
+      focusElement(firstAction);
+    }
+  };
+
+  menuActionsKeydown = (event: KeyboardEvent): void => {
+    const { key, target } = event;
+
+    if (!this.isValidKey(key, SUPPORTED_ARROW_KEYS)) {
+      return;
+    }
+
+    const actions = this.queryActions();
+    const { length } = actions;
+    const currentIndex = actions.indexOf(target as HTMLCalciteActionElement);
+
+    if (!length || currentIndex === -1) {
+      return;
+    }
+
+    event.preventDefault();
+
+    if (key === "ArrowUp") {
+      const value = getRoundRobinIndex(currentIndex - 1, length);
+      const previousAction = actions[value];
+      focusElement(previousAction);
+    }
+
+    if (key === "ArrowDown") {
+      const value = getRoundRobinIndex(currentIndex + 1, length);
+      const nextAction = actions[value];
+      focusElement(nextAction);
+    }
+  };
+
+  menuActionsContainerKeyDown = (event: KeyboardEvent): void => {
+    const { key } = event;
+
+    if (key === "Escape") {
+      this.menuOpen = false;
+    }
   };
 
   // --------------------------------------------------------------------------
@@ -117,18 +207,18 @@ export class CalciteFlowItem {
   renderBackButton(rtl: boolean): VNode {
     const { showBackButton, textBack, backButtonClick } = this;
 
-    const path = rtl ? chevronRight16F : chevronLeft16F;
+    const icon = rtl ? ICONS.backRight : ICONS.backLeft;
 
     return showBackButton ? (
       <calcite-action
-        slot="header-leading-content"
+        slot={SLOTS.headerLeadingContent}
         key="back-button"
         aria-label={textBack}
         text={textBack}
         class={CSS.backButton}
         onClick={backButtonClick}
       >
-        <CalciteIcon size="16" path={path} />
+        <calcite-icon scale="s" filled icon={icon} />
       </calcite-action>
     ) : null;
   }
@@ -144,8 +234,9 @@ export class CalciteFlowItem {
         aria-label={menuLabel}
         text={menuLabel}
         onClick={this.toggleMenuOpen}
+        onKeyDown={this.menuButtonKeyDown}
       >
-        <CalciteIcon size="16" path={ellipsis16} />
+        <calcite-icon scale="s" icon={ICONS.menu} />
       </calcite-action>
     );
   }
@@ -154,18 +245,21 @@ export class CalciteFlowItem {
     const { menuOpen } = this;
 
     return (
-      <div class={classnames(CSS.menu, { [CSS.menuOpen]: menuOpen })}>
-        <slot name="menu-actions" />
+      <div
+        class={classnames(CSS.menu, { [CSS.menuOpen]: menuOpen })}
+        onKeyDown={this.menuActionsKeydown}
+      >
+        <slot name={SLOTS.menuActions} />
       </div>
     );
   }
 
   renderFooterActions(): VNode {
-    const hasFooterActions = !!this.el.querySelector("[slot=footer-actions]");
+    const hasFooterActions = !!this.el.querySelector(`[slot=${SLOTS.footerActions}]`);
 
     return hasFooterActions ? (
       <div slot="footer" class={CSS.footerActions}>
-        <slot name="footer-actions" />
+        <slot name={SLOTS.footerActions} />
       </div>
     ) : null;
   }
@@ -173,14 +267,14 @@ export class CalciteFlowItem {
   renderSingleActionContainer(): VNode {
     return (
       <div class={CSS.singleActionContainer}>
-        <slot name="menu-actions" />
+        <slot name={SLOTS.menuActions} />
       </div>
     );
   }
 
   renderMenuActionsContainer(): VNode {
     return (
-      <div class={CSS.menuContainer}>
+      <div class={CSS.menuContainer} onKeyDown={this.menuActionsContainerKeyDown}>
         {this.renderMenuButton()}
         {this.renderMenuActions()}
       </div>
@@ -188,7 +282,7 @@ export class CalciteFlowItem {
   }
 
   renderHeaderActions(): VNode {
-    const menuActionsNode = this.el.querySelector("[slot=menu-actions]");
+    const menuActionsNode = this.el.querySelector(`[slot=${SLOTS.menuActions}]`);
 
     const hasMenuActionsInBlacklisted =
       menuActionsNode && menuActionsNode.closest(BLACKLISTED_MENU_ACTIONS_COMPONENTS.join(","));
@@ -203,16 +297,39 @@ export class CalciteFlowItem {
         ? this.renderMenuActionsContainer()
         : null;
 
-    return menuActionsNodes ? <div slot="header-trailing-content">{menuActionsNodes}</div> : null;
+    return menuActionsNodes ? (
+      <div slot={SLOTS.headerTrailingContent} class={CSS.headerActions}>
+        {menuActionsNodes}
+      </div>
+    ) : null;
   }
 
   renderHeading(): VNode {
     const { heading } = this;
 
     return heading ? (
-      <h2 class={CSS.heading} slot="header-content">
+      <h2 class={CSS.heading} slot={SLOTS.headerContent}>
         {heading}
       </h2>
+    ) : null;
+  }
+
+  renderSummary(): VNode {
+    const { summary } = this;
+
+    return summary ? <span class={CSS.summary}>{summary}</span> : null;
+  }
+
+  renderHeader(): VNode {
+    const headingNode = this.renderHeading();
+
+    const summaryNode = this.renderSummary();
+
+    return headingNode || summaryNode ? (
+      <header class={CSS.header} slot={SLOTS.headerContent}>
+        {headingNode}
+        {summaryNode}
+      </header>
     ) : null;
   }
 
@@ -223,9 +340,19 @@ export class CalciteFlowItem {
 
     return (
       <Host>
-        <calcite-panel loading={this.loading} disabled={this.disabled}>
+        <calcite-panel
+          loading={this.loading}
+          disabled={this.disabled}
+          height-scale={this.heightScale}
+          class={classnames({
+            [CSS_UTILITY.rtl]: rtl
+          })}
+        >
           {this.renderBackButton(rtl)}
-          {this.renderHeading()}
+          <div class={CSS.header} slot="header-content">
+            {this.renderHeading()}
+            {this.renderSummary()}
+          </div>
           {this.renderHeaderActions()}
           <slot />
           {this.renderFooterActions()}