+++ src/components/calcite-flow-item/resources.ts
@@ -1,9 +1,12 @@
 export const BLACKLISTED_MENU_ACTIONS_COMPONENTS = ["calcite-pick-list", "calcite-value-list"];
 
 export const CSS = {
+  header: "header-content",
   heading: "heading",
+  summary: "summary",
   backButton: "back-button",
   footerActions: "footer-actions",
+  headerActions: "header-actions",
   singleActionContainer: "single-action-container",
   menuContainer: "menu-container",
   menuButton: "menu-button",
@@ -15,4 +18,18 @@ export const TEXT = {
   back: "Back",
   open: "Open",
   close: "Close"
+};
+
+export const SLOTS = {
+  headerContent: "header-content",
+  headerLeadingContent: "header-leading-content",
+  headerTrailingContent: "header-trailing-content",
+  menuActions: "menu-actions",
+  footerActions: "footer-actions"
+};
+
+export const ICONS = {
+  menu: "ellipsis",
+  backLeft: "chevron-left",
+  backRight: "chevron-right"
 };