+++ src/components/calcite-panel/resources.ts
@@ -1,3 +1,5 @@
+import { CalciteScale } from "../interfaces";
+
 export const CSS = {
   container: "container",
   header: "header",
@@ -10,4 +12,17 @@ export const CSS = {
 
 export const TEXT = {
   close: "Close"
+};
+
+export const SCALES: CalciteScale[] = ["s", "m", "l"];
+
+export const ICONS = {
+  close: "x"
+};
+
+export const SLOTS = {
+  headerContent: "header-content",
+  headerLeadingContent: "header-leading-content",
+  headerTrailingContent: "header-trailing-content",
+  footer: "footer"
 };